Refactored DefaultProtocol
Refactored DefaultProtocol by creating GenericUtilsProtocol
