+++ testar/src/org/fruit/monkey/DefaultProtocol.java
@@ -2031,143 +2031,5 @@ public class DefaultProtocol extends RuntimeControlsProtocol {
 		return false;
 	}
 
-	/**
-     * Adds sliding actions (like scroll, drag and drop) to the given Set of Actions
-     * @param actions
-     * @param ac
-     * @param scrollArrowSize
-     * @param scrollThick
-     * @param widget
-     */
-    protected void addSlidingActions(Set<Action> actions, StdActionCompiler ac, double scrollArrowSize, double scrollThick, Widget widget, State state){
-        Drag[] drags = null;
-        //If there are scroll (drags/drops) actions possible
-        if((drags = widget.scrollDrags(scrollArrowSize,scrollThick)) != null){
-            //For each possible drag, create an action and add it to the derived actions
-            for (Drag drag : drags){
-                //Create a slide action with the Action Compiler, and add it to the set of derived actions
-                actions.add(ac.slideFromTo(
-                        new AbsolutePosition(Point.from(drag.getFromX(),drag.getFromY())),
-                        new AbsolutePosition(Point.from(drag.getToX(),drag.getToY())),
-                        widget
-                ));
-
-            }
-        }
-    }
-
-	/**
-	 * This method waits until the widget with a matching Tag value (case sensitive) is found or the retry limit is reached.
-	 * If a matching widget is found, left mouse button is clicked on it and return value is true.
-	 * Else returns false
-	 *
-	 * @param tag for example: org.fruit.alayer.Tags.Title
-	 * @param value
-	 * @param state
-	 * @param system needed for updating the state between retries
-	 * @param maxNumberOfRetries int number of times
-	 * @param waitBetween double in seconds
-	 * @return
-	 */
-	protected boolean waitAndLeftClickWidgetWithMatchingTag(Tag tag, String value, State state, SUT system, int maxNumberOfRetries, double waitBetween){
-		int numberOfRetries = 0;
-		while(numberOfRetries<maxNumberOfRetries){
-			//looking for a widget with matching tag value:
-			Widget widget = getWidgetWithMatchingTag(tag,value,state);
-			if(widget!=null){
-				StdActionCompiler ac = new AnnotatingActionCompiler();
-				//System.out.println("DEBUG: left mouse click on a widget with "+tag.toString()+"=" + value);
-				executeAction(system,state,ac.leftClickAt(widget));
-				// is waiting needed after the action has been executed?
-				return true;
-			}
-			else{
-				Util.pause(waitBetween);
-				state = getState(system);
-				numberOfRetries++;
-			}
-		}
-		System.out.println("Matching widget was not found, "+tag.toString()+"=" + value);
-		printTagValuesOfWidgets(tag,state);
-		return false;
-	}
-
-	/**
-	 * This method waits until the widget with a matching Tag value (case sensitive) is found or the retry limit is reached.
-	 * If a matching widget is found, left mouse button is clicked on it, the given text is typed into it, and return value is true.
-	 * Else returns false
-	 *
-	 * @param tag for example: org.fruit.alayer.Tags.Title
-	 * @param value
-	 * @param textToType types the given text by replacing the existing text
-	 * @param state
-	 * @param system needed for updating the state between retries
-	 * @param maxNumberOfRetries int number of times
-	 * @param waitBetween double in seconds
-	 * @return
-	 */
-	protected boolean waitLeftClickAndTypeIntoWidgetWithMatchingTag(Tag tag, String value, String textToType, State state, SUT system, int maxNumberOfRetries, double waitBetween){
-		int numberOfRetries = 0;
-		while(numberOfRetries<maxNumberOfRetries){
-			//looking for a widget with matching tag value:
-			Widget widget = getWidgetWithMatchingTag(tag,value,state);
-			if(widget!=null){
-				StdActionCompiler ac = new AnnotatingActionCompiler();
-				executeAction(system,state,ac.clickTypeInto(widget, textToType, true));
-				// is waiting needed after the action has been executed?
-				return true;
-			}
-			else{
-				Util.pause(waitBetween);
-				state = getState(system);
-				numberOfRetries++;
-			}
-		}
-		System.out.println("Matching widget was not found, "+tag.toString()+"=" + value);
-		printTagValuesOfWidgets(tag,state);
-		return false;
-	}
-
-
-	/**
-	 * Iterates the widgets of the state until a widget with matching tag value is found.
-	 * The value is case sensitive.
-	 *
-	 * @param tag
-	 * @param value
-	 * @param state
-	 * @return the matching widget if found, null if not found
-	 */
-	protected Widget getWidgetWithMatchingTag(Tag tag, String value, State state){
-		for(Widget widget:state){
-			if(widget.get(tag, null)==null){
-				// this widget did not have a value for the given tag
-			}
-			else if(widget.get(tag, null).toString().equals(value)){
-				return widget;
-			}
-			else if(widget.get(tag, null).toString().contains(value)) {
-				return widget;
-			}
-		}
-		return null;
-	}
-
-	/**
-	 * Prints to system out all the widgets that have some value in the given tag.
-	 *
-	 * @param tag
-	 * @param state
-	 */
-	protected void printTagValuesOfWidgets(Tag tag, State state){
-		for(Widget widget:state){
-			if(widget.get(tag, null)==null){
-				// this widget did not have a value for the given tag
-			}
-			else{
-				System.out.println(tag.toString()+"=" + widget.get(tag, null).toString()+ "; Description of the widget="+widget.get(Tags.Desc, ""));
-			}
-		}
-	}
 
 }
