+++ testar/src/org/testar/protocols/DesktopProtocol.java
@@ -47,7 +47,7 @@ import java.util.Set;
 import static org.fruit.alayer.Tags.Blocked;
 import static org.fruit.alayer.Tags.Enabled;
 
-public class DesktopProtocol extends ClickFilterLayerProtocol {
+public class DesktopProtocol extends GenericUtilsProtocol {
     //Attributes for adding slide actions
     protected static double SCROLL_ARROW_SIZE = 36; // sliding arrows
     protected static double SCROLL_THICK = 16; //scroll thickness
